old-interations/v3_batchprocessing/requirements.txt
# requirements.txt
# Requirements for SeamlessM4T-based translation system (HuggingFace version)

# Core ML frameworks
torch>=2.0.0
torchaudio>=2.0.0

# HuggingFace Transformers (includes SeamlessM4T)
transformers>=4.35.0
sentencepiece
accelerate
scipy

# Speaker diarization
pyannote.audio
huggingface-hub

# Audio processing
pydub
soundfile
moviepy==1.0.3

# Utilities
python-dotenv
numpy

# Optional: For better audio device support
sounddevice

openai-whisper
bark
nltk
librosa